fix(core): moving some files to a shared location
Part of ENG-433
